Page now works correctly
Discovered the source of the problem of why console.log was not working. In the original html the path to the js file started with type rather than src. Stress from doing my first hackathon preventing me from recognizing the error at the time and completing the project in time. This will not happen again.

On the bright side, I now know how to turn form data into a a json object. The next step will be to learn how to post the json object into a json file for future use.